+++ embedded/android/PhoneController/app/src/main/java/com/ndomx/phonecontroller/ui/components/ConnectButton.kt
@@ -1,8 +1,7 @@
-package com.ndomx.phonecontroller.ui
+package com.ndomx.phonecontroller.ui.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.size
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
