+++ embedded/android/PhoneController/app/src/main/java/com/ndomx/phonecontroller/ui/components/ConnectionStatusIndicator.kt
@@ -1,4 +1,4 @@
-package com.ndomx.phonecontroller.ui
+package com.ndomx.phonecontroller.ui.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
