+++ embedded/android/PhoneController/app/src/main/java/com/ndomx/phonecontroller/ui/components/SensitiveText.kt
@@ -1,9 +1,7 @@
-package com.ndomx.phonecontroller.ui
+package com.ndomx.phonecontroller.ui.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Lock
-import androidx.compose.material.icons.filled.ThumbUp
 import androidx.compose.material.icons.filled.Visibility
 import androidx.compose.material.icons.filled.VisibilityOff
 import androidx.compose.material3.Icon
